pyproject.toml
[tool.poetry]
name = "kafka-spark orchestrated with airflow"
version = "0.1.0"
description = "project_description"
authors = ["Volodymyr Tseluiko <vladimir.tseluyko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.1"
pyspark = "^3.1.1"
py-postgresql = "^1.2.2"
cassandra-driver = "^3.9.0"
autopep8 = "^1.5.7"
essential-generators = "^1.0"
requests = "^2.2.6.0"
faust = "^1.10.4"
importlib-resources = "^5.4.0"
SQLAlchemy = "1.3.24"
apache-airflow = "^2.2.2"
pytest = "6.2.5"

[tool.poetry.dev-dependencies]
pyspark = "^3.1.1"
py-postgresql = "^1.2.2"
flask = "^1.1.2"

cassandra-driver = "^3.9.0"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
